requirements.txt
streamlit>=1.32
python-dotenv>=1.0.1
pydantic>=2.7
openai>=1.40.0
langgraph>=0.2.40
langchain-core>=0.3.0
tiktoken>=0.7.0

# Parsing
python-docx>=1.1.2
PyMuPDF>=1.24.7

# RAG
faiss-cpu>=1.8.0.post1
numpy>=1.26

# PPTX rendering
python-pptx>=0.6.23

# Utilities
rich>=13.7
